+++ R/items_pilot.R
@@ -2,7 +2,7 @@
 #############################################################################
 #' Small simulated item pool example.
 #'
-#' A \code{data.frame} containing 100 uncalibrated items with different categorical and metric properties.
+#' A \code{data.frame} containing 100 not yet calibrated items with different categorical and metric properties.
 #'
 #' @format A \code{data.frame} .
 #'\describe{